+++ angular-client/src/app/models/user.ts
@@ -1,0 +1,10 @@
+export class User {
+
+    public givenName: string;
+    public familyName: string;
+
+    public email: string;
+
+    public groups: string[];
+
+}