+++ angular-client/src/app/services/auth.guard.ts
@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Router, CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
-
-import { OidcSecurityService } from 'angular-auth-oidc-client';
+import { map } from 'rxjs/operators';
 
 import { AuthService } from './auth.service';
 
@@ -14,23 +12,23 @@ import { AuthService } from './auth.service';
 
     constructor(
         private router: Router,
-        private oidcSecurityService: OidcSecurityService,
         private authService: AuthService) { }
 
     public canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> | boolean {
-        return this.oidcSecurityService.getIsAuthorized()
-            .map((isAuthorized: boolean) => {
-                if (isAuthorized) {
+        return this.authService.isSignedIn().pipe(
+            map((signedIn: boolean) => {
+                if (signedIn) {
                     return true;
                 }
 
                 // Stores the attempted URL for redirecting.
-                this.authService.setRedirectUrl(state.url);
+                this.authService.setItem('redirectUrl', state.url);
 
                 // Not signed in so redirects to unauthorized page.
                 this.router.navigate(['/unauthorized']);
                 return false;
-            });
+            })
+        );
     }
 
 }