+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Angular OIDC PHP - Changelog
 
+<a name="Nov 28, 2017"></a>
+### Nov 28, 2017
+* Switch to angular-oauth2-oidc
+
 <a name="Nov 11, 2017"></a>
 ### Nov 11, 2017
 * Upgrade to Angular v5 & Angular CLI 1.5